+++ src/pages/home/components/learnMoreBtn.jsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { Button } from '@mui/material';
+import PropTypes from 'prop-types';
+import styled from 'styled-components';
+
+function LearnMoreBtn({ href }) {
+  return (
+    <Wrapper>
+      <StyledButton
+        className="LearnMoreBtn"
+        variant="contained"
+        color="primary"
+        href={href}
+        style={{ backgroundColor: 'rgb(102, 0, 0)' }}
+      >
+        Learn More
+      </StyledButton>
+    </Wrapper>
+  );
+}
+
+const Wrapper = styled.div`
+  padding-top: 2rem;
+`;
+
+const StyledButton = styled(Button)`
+  font-size: 14px;
+  display: block;
+  align-items: center;
+  justify-content: center;
+  width: 200px;
+  padding-top: 10vh;
+`;
+
+LearnMoreBtn.propTypes = {
+  href: PropTypes.string,
+};
+
+LearnMoreBtn.defaultProps = {
+  href: '/',
+};
+
+export default LearnMoreBtn;
