+++ src/pages/home/components/whoWeAre.jsx
@@ -36,7 +36,7 @@ function WhoWeAre() {
 
 const Wrapper = styled.div`
   text-align: left;
-  padding: 10vh 10vw;
+  padding: 15vh 10vw;
   background-color: white;
   @media (max-width: 960px) {
     text-align: center;
