+++ src/pages/home/components/whyWeDoIt.jsx
@@ -30,7 +30,7 @@ function WhyWeDoIt() {
 
 const Wrapper = styled.div`
   text-align: left;
-  padding: 10vh 10vw;
+  padding: 15vh 10vw;
   background-color: white;
   @media (max-width: 960px) {
     text-align: center;
